+++ test/accesstoken_dio_test.dart
@@ -1,5 +1,4 @@
 import 'package:dio/dio.dart';
-import 'package:dio_http_formatter/dio_http_formatter.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:supertokens_flutter/src/dio-interceptor-wrapper.dart';
@@ -28,8 +27,8 @@ void main() {
     Dio dio = Dio(
       BaseOptions(
         baseUrl: apiBasePath,
-        connectTimeout: 5000,
-        receiveTimeout: 500,
+        connectTimeout: Duration(seconds: 5),
+        receiveTimeout: Duration(milliseconds: 500),
       ),
     );
     dio.interceptors.add(SuperTokensInterceptorWrapper(client: dio));