+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-11-13
+
+- Added Dio extension for interceptor setup
+- Updated the mutex package to version ^3.1.0
+
+### Breaking Changes
+
+- Updated the dio package to version ^5.0.0
+
 ## [0.2.8] - 2023-09-13
 
 - Adds 1.18 to the list of supported FDI versions